requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
qiskit==0.45.0
qiskit-ibm-runtime==0.15.1
qiskit-ibm-provider==0.7.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-multipart==0.0.6
websockets==12.0
apscheduler==3.10.4
redis==5.0.1
celery==5.3.4
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
